Extract call formatting out of KotlinInputAstVisitor into CallFormatter
Purely mechanical change that extracts all `visit*` methods related to call expressions into corresponding `format*`
