+++ core/src/main/kotlin/org/jetbrains/ktfmt/format/KotlinInputAstVisitor.kt
@@ -29,11 +29,9 @@ import com.google.googlejavaformat.Output.BreakTag
 import java.util.ArrayDeque
 import java.util.Optional
 import kotlin.jvm.optionals.getOrNull
-import org.jetbrains.kotlin.com.intellij.psi.PsiComment
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.com.intellij.psi.stubs.PsiFileStubImpl
 import org.jetbrains.kotlin.lexer.KtTokens
-import org.jetbrains.kotlin.psi.KtArrayAccessExpression
 import org.jetbrains.kotlin.psi.KtBackingField
 import org.jetbrains.kotlin.psi.KtBinaryExpression
 import org.jetbrains.kotlin.psi.KtBinaryExpressionWithTypeRHS
@@ -57,7 +55,6 @@ import org.jetbrains.kotlin.psi.KtDelegatedSuperTypeEntry
 import org.jetbrains.kotlin.psi.KtDestructuringDeclaration
 import org.jetbrains.kotlin.psi.KtDestructuringDeclarationEntry
 import org.jetbrains.kotlin.psi.KtDoWhileExpression
-import org.jetbrains.kotlin.psi.KtDotQualifiedExpression
 import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.KtEnumEntry
 import org.jetbrains.kotlin.psi.KtExpression
@@ -68,7 +65,6 @@ import org.jetbrains.kotlin.psi.KtImportDirective
 import org.jetbrains.kotlin.psi.KtIsExpression
 import org.jetbrains.kotlin.psi.KtLabelReferenceExpression
 import org.jetbrains.kotlin.psi.KtLabeledExpression
-import org.jetbrains.kotlin.psi.KtLambdaArgument
 import org.jetbrains.kotlin.psi.KtLambdaExpression
 import org.jetbrains.kotlin.psi.KtModifierList
 import org.jetbrains.kotlin.psi.KtNamedFunction
@@ -94,43 +90,34 @@ import org.jetbrains.kotlin.psi.KtThisExpression
 import org.jetbrains.kotlin.psi.KtThrowExpression
 import org.jetbrains.kotlin.psi.KtTryExpression
 import org.jetbrains.kotlin.psi.KtTypeAlias
-import org.jetbrains.kotlin.psi.KtTypeArgumentList
 import org.jetbrains.kotlin.psi.KtTypeConstraintList
 import org.jetbrains.kotlin.psi.KtTypeParameterList
 import org.jetbrains.kotlin.psi.KtTypeReference
 import org.jetbrains.kotlin.psi.KtValueArgument
-import org.jetbrains.kotlin.psi.KtValueArgumentList
 import org.jetbrains.kotlin.psi.KtWhenConditionInRange
 import org.jetbrains.kotlin.psi.KtWhenConditionIsPattern
 import org.jetbrains.kotlin.psi.KtWhenConditionWithExpression
 import org.jetbrains.kotlin.psi.KtWhenExpression
 import org.jetbrains.kotlin.psi.KtWhileExpression
-import org.jetbrains.kotlin.psi.psiUtil.children
 import org.jetbrains.kotlin.psi.psiUtil.startOffset
-import org.jetbrains.kotlin.psi.psiUtil.startsWithComment
 import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes
 import org.jetbrains.kotlin.psi.stubs.impl.KotlinPlaceHolderStubImpl
 import org.jetbrains.ktfmt.format.visitor.AbstractFormatterVisitor
 import org.jetbrains.ktfmt.format.visitor.AnnotationFormatter
+import org.jetbrains.ktfmt.format.visitor.CallFormatter
 import org.jetbrains.ktfmt.format.visitor.ExpressionFormatter
 import org.jetbrains.ktfmt.format.visitor.FileFormatter
 import org.jetbrains.ktfmt.format.visitor.ListFormatter
 import org.jetbrains.ktfmt.format.visitor.TypeFormatter
 import org.jetbrains.ktfmt.format.visitor.asIndent
 import org.jetbrains.ktfmt.format.visitor.block
-import org.jetbrains.ktfmt.format.visitor.chainParts
-import org.jetbrains.ktfmt.format.visitor.chainRoot
-import org.jetbrains.ktfmt.format.visitor.chainedSelectorsHaveValueArguments
 import org.jetbrains.ktfmt.format.visitor.fenceComments
 import org.jetbrains.ktfmt.format.visitor.hasEmptyParenthesis
 import org.jetbrains.ktfmt.format.visitor.hasLineBreakingCommentBefore
-import org.jetbrains.ktfmt.format.visitor.hasSourceNewlineInLambdaBody
 import org.jetbrains.ktfmt.format.visitor.isBlockLikeCall
 import org.jetbrains.ktfmt.format.visitor.isChainedBlockLikeCall
 import org.jetbrains.ktfmt.format.visitor.isChainedScopingFunction
-import org.jetbrains.ktfmt.format.visitor.isLambda
 import org.jetbrains.ktfmt.format.visitor.isLambdaOrScopingFunction
-import org.jetbrains.ktfmt.format.visitor.isMultilineScopingFunction
 import org.jetbrains.ktfmt.format.visitor.sync
 import org.jetbrains.ktfmt.format.visitor.token
 import org.jetbrains.ktfmt.util.CONTEXT_PARAMETER_LIST
@@ -143,13 +130,14 @@ open class KotlinInputAstVisitor(
 ) :
     AbstractFormatterVisitor(),
     AnnotationFormatter,
+    CallFormatter,
     ExpressionFormatter,
     FileFormatter,
     ListFormatter,
     TypeFormatter {
 
   /** Standard indentation for a block */
-  private val blockIndent: Indent.Const = options.blockIndent.asIndent
+  override val blockIndent: Indent.Const = options.blockIndent.asIndent
 
   /**
    * Standard indentation for a long expression or function call, it is different than block
@@ -157,10 +145,10 @@ open class KotlinInputAstVisitor(
    */
   override val expressionBreakIndent: Indent.Const = options.continuationIndent.asIndent
 
-  private val blockPlusExpressionBreakIndent: Indent.Const =
+  override val blockPlusExpressionBreakIndent: Indent.Const =
       (options.blockIndent + options.continuationIndent).asIndent
 
-  private val doubleExpressionBreakIndent: Indent.Const = (options.continuationIndent * 2).asIndent
+  override val doubleExpressionBreakIndent: Indent.Const = (options.continuationIndent * 2).asIndent
 
   override val expressionBreakNegativeIndent: Indent.Const = (-options.continuationIndent).asIndent
 
@@ -168,7 +156,7 @@ open class KotlinInputAstVisitor(
   override val inExpression = ArrayDeque(ImmutableList.of(false))
 
   /** Tracks whether we are handling an import directive */
-  private var inImport = false
+  override var inImport = false
 
   /** Example: `fun foo(n: Int) { println(n) }` */
   override fun visitNamedFunction(function: KtNamedFunction) {
@@ -319,26 +307,6 @@ open class KotlinInputAstVisitor(
     builder.token("}", blockIndent)
   }
 
-  private fun visitStatement(statement: PsiElement) {
-    builder.block(ZERO) { visit(statement) }
-    builder.guessToken(";")
-  }
-
-  private fun visitStatements(statements: Array<PsiElement>) {
-    var first = true
-    builder.guessToken(";")
-    for (statement in statements) {
-      builder.forcedBreak()
-      if (!first) {
-        builder.blankLineWanted(OpsBuilder.BlankLineWanted.PRESERVE)
-      }
-      first = false
-      markForPartialFormat()
-      visitStatement(statement)
-      markForPartialFormat()
-    }
-  }
-
   override fun visitProperty(property: KtProperty) {
     builder.sync(property)
     builder.block(ZERO) {
@@ -363,474 +331,15 @@ open class KotlinInputAstVisitor(
     }
   }
 
-  /**
-   * Example: "com.facebook.bla.bla" in imports or "a.b.c.d" in expressions.
-   *
-   * There's a few cases that are different. We deal with imports by keeping them on the same line.
-   * For regular chained expressions we go the left most descendant so we can start indentation only
-   * before the first break (a `.` or `?.`), and keep the seem indentation for this chain of calls.
-   */
-  override fun visitQualifiedExpression(expression: KtQualifiedExpression) {
-    builder.sync(expression)
-    val receiver = expression.receiverExpression
-    when {
-      inImport -> {
-        visit(receiver)
-        val selectorExpression = expression.selectorExpression
-        if (selectorExpression != null) {
-          builder.token(".")
-          visit(selectorExpression)
-        }
-      }
-      receiver is KtStringTemplateExpression -> {
-        builder.block(expressionBreakIndent) {
-          visit(receiver)
-          builder.breakOp(Doc.FillMode.UNIFIED, "", ZERO)
-          builder.token(expression.operationSign.value)
-          visit(expression.selectorExpression)
-        }
-      }
-      receiver is KtWhenExpression -> {
-        builder.block(ZERO) {
-          visit(receiver)
-          builder.token(expression.operationSign.value)
-          visit(expression.selectorExpression)
-        }
-      }
-      expression.isChainedScopingFunction &&
-          expression.chainRoot.isMultilineScopingFunction &&
-          !chainedSelectorsHaveValueArguments(expression) -> {
-        visitChainedScopingFunction(expression, emitLeadingBreak = false)
-      }
-      expression.isChainedBlockLikeCall -> {
-        visitChainedBlockLikeCall(expression, emitLeadingBreak = false)
-      }
-      else -> {
-        emitQualifiedExpression(expression)
-      }
-    }
-  }
-
-  /** Extra data to help [emitQualifiedExpression] know when to open and close a group */
-  private class GroupingInfo {
-    var groupOpenCount = 0
-    var shouldCloseGroup = false
-  }
-
-  /**
-   * Handles a chain of qualified expressions, i.e. `a[5].b!!.c()[4].f()`
-   *
-   * This is by far the most complicated part of this formatter. We start by breaking the expression
-   * to the steps it is executed in (which are in the opposite order of how the syntax tree is
-   * built).
-   *
-   * We then calculate information to know which parts need to be groups, and finally go part by
-   * part, emitting it to the [builder] while closing and opening groups.
-   */
-  private fun emitQualifiedExpression(expression: KtExpression) {
-    val parts = expression.chainParts
-    // whether we want to make a lambda look like a block, this make Kotlin DSLs look as expected
-    val useBlockLikeLambdaStyle = parts.last().isLambda && parts.count { it.isLambda } == 1
-    val groupingInfos = computeGroupingInfo(parts, useBlockLikeLambdaStyle)
-    builder.block(expressionBreakIndent) {
-      val nameTag = genSym() // allows adjusting arguments indentation if a break will be made
-      for ((index, ktExpression) in parts.withIndex()) {
-        if (ktExpression is KtQualifiedExpression) {
-          builder.breakOp(Doc.FillMode.UNIFIED, "", ZERO, Optional.of(nameTag))
-        }
-        repeat(groupingInfos[index].groupOpenCount) { builder.open(ZERO) }
-        when (ktExpression) {
-          is KtQualifiedExpression -> {
-            builder.token(ktExpression.operationSign.value)
-            val selectorExpression = ktExpression.selectorExpression
-            if (selectorExpression !is KtCallExpression) {
-              // selector is a simple field access
-              visit(selectorExpression)
-              if (groupingInfos[index].shouldCloseGroup) {
-                builder.close()
-              }
-            } else {
-              // selector is a function call, we may close a group after its name
-              // emit `doIt` from `doIt(1, 2) { it }`
-              visit(selectorExpression.calleeExpression)
-              // close groups according to instructions
-              if (groupingInfos[index].shouldCloseGroup) {
-                builder.close()
-              }
-              // close group due to last lambda to allow block-like style in `as.forEach { ... }`
-              val isTrailingLambda = useBlockLikeLambdaStyle && index == parts.size - 1
-              if (isTrailingLambda) {
-                builder.close()
-              }
-              // A block-like (exploded) selector call is laid out like the last part: its
-              // arguments are indented once relative to the call itself, and its closing paren
-              // returns to the call's indent, even when chained selectors follow it. This only
-              // applies when trailing commas are preserved (the block-like style); when ktfmt
-              // manages trailing commas, exploded chained calls keep the regular extra indent.
-              val isLastPartOrBlockLikeCall =
-                  index == parts.size - 1 ||
-                      !options.manageTrailingCommas && selectorExpression.isBlockLikeCall
-              val argsIndentElse = if (isLastPartOrBlockLikeCall) ZERO else expressionBreakIndent
-              val lambdaIndentElse = if (isTrailingLambda) expressionBreakNegativeIndent else ZERO
-              val negativeLambdaIndentElse = if (isTrailingLambda) expressionBreakIndent else ZERO
-
-              // emit `(1, 2) { it }` from `doIt(1, 2) { it }`
-              visitCallElement(
-                  null,
-                  selectorExpression.typeArgumentList,
-                  selectorExpression.valueArgumentList,
-                  selectorExpression.lambdaArguments,
-                  argumentsIndent = Indent.If.make(nameTag, expressionBreakIndent, argsIndentElse),
-                  lambdaIndent = Indent.If.make(nameTag, ZERO, lambdaIndentElse),
-                  negativeLambdaIndent = Indent.If.make(nameTag, ZERO, negativeLambdaIndentElse),
-              )
-            }
-          }
-          is KtArrayAccessExpression -> {
-            visitArrayAccessBrackets(ktExpression)
-            builder.close()
-          }
-          is KtPostfixExpression -> {
-            builder.token(ktExpression.operationReference.text)
-            builder.close()
-          }
-          else -> {
-            check(index == 0)
-            visit(ktExpression)
-          }
-        }
-      }
-    }
-  }
-
-  /**
-   * Generates the [GroupingInfo] array to go with an array of [KtQualifiedExpression] parts
-   *
-   * For example, the expression `a.b[2].c.d()` is made of four expressions:
-   * 1. [KtQualifiedExpression] `a.b[2].c . d()` (this will be `parts[4]`)
-   * 1. [KtQualifiedExpression] `a.b[2] . c` (this will be `parts[3]`)
-   * 2. [KtArrayAccessExpression] `a.b [2]` (this will be `parts[2]`)
-   * 3. [KtQualifiedExpression] `a . b` (this will be `parts[1]`)
-   * 4. [KtSimpleNameExpression] `a` (this will be `parts[0]`)
-   *
-   * Once in parts, these are in the reverse order. To render the array correct we need to make sure
-   * `b` and [2] are in a group so we avoid splitting them. To do so we need to open a group for `b`
-   * (that will be done in part 2), and always close a group for an array.
-   *
-   * Here is the same expression, with justified braces marking the groupings it will get:
-   * ```
-   *  a . b [2] . c . d ()
-   * {a . b} --> Grouping `a.b` because it can be a package name or simple field access so we add 1
-   *             to the number of groups to open at groupingInfos[0], and mark to close a group at
-   *             groupingInfos[1]
-   * {a . b [2]} --> Grouping `a.b` with `[2]`, since otherwise we may break inside the brackets
-   *                 instead of preferring breaks before dots. So we open a group at [0], but since
-   *                 we always close a group after brackets, we don't store that information.
-   *             {c . d} --> another group to attach the first function name to the fields before it
-   *                         this time we don't start the group in the beginning, and use
-   *                         lastIndexToOpen to track the spot after the last time we stopped
-   *                         grouping.
-   * ```
-   *
-   * The final expression with groupings:
-   * ```
-   * {{a.b}[2]}.{c.d}()
-   * ```
-   */
-  private fun computeGroupingInfo(
-      parts: List<KtExpression>,
-      useBlockLikeLambdaStyle: Boolean,
-  ): List<GroupingInfo> {
-    val groupingInfos = List(parts.size) { GroupingInfo() }
-    var lastIndexToOpen = 0
-    for ((index, part) in parts.withIndex()) {
-      when (part) {
-        is KtQualifiedExpression -> {
-          val receiverExpression = part.receiverExpression
-          val previous =
-              (receiverExpression as? KtQualifiedExpression)?.selectorExpression
-                  ?: receiverExpression
-          val current = checkNotNull(part.selectorExpression)
-          if (
-              lastIndexToOpen == 0 &&
-                  shouldGroupPartWithPrevious(parts, part, index, previous, current)
-          ) {
-            // this and the previous items should be grouped for better style
-            // we add another group to open in index 0
-            groupingInfos[0].groupOpenCount++
-            // we don't always close a group when emitting this node, so we need this flag to
-            // mark if we need to close a group
-            groupingInfos[index].shouldCloseGroup = true
-          } else {
-            // use this index in to open future groups for arrays and postfixes
-            // we will also stop grouping field access to the beginning of the expression
-            lastIndexToOpen = index
-          }
-        }
-        is KtArrayAccessExpression,
-        is KtPostfixExpression -> {
-          // we group these with the last item with a name, and we always close them
-          groupingInfos[lastIndexToOpen].groupOpenCount++
-        }
-      }
-    }
-    if (useBlockLikeLambdaStyle) {
-      // a trailing lambda adds a group that we stop before emitting the lambda
-      groupingInfos[0].groupOpenCount++
-    }
-    return groupingInfos
-  }
-
-  /** Decide whether a [KtQualifiedExpression] part should be grouped with the previous part */
-  private fun shouldGroupPartWithPrevious(
-      parts: List<KtExpression>,
-      part: KtExpression,
-      index: Int,
-      previous: KtExpression,
-      current: KtExpression,
-  ): Boolean {
-    // this is the second, and the first is short, avoid `.` "hanging in air"
-    if (index == 1 && previous.text.length < options.continuationIndent) {
-      return true
-    }
-    // the previous part is `this` or `super`
-    if (previous is KtSuperExpression || previous is KtThisExpression) {
-      return true
-    }
-    // this and the previous part are a package name, type name, or property
-    if (
-        previous is KtSimpleNameExpression &&
-            current is KtSimpleNameExpression &&
-            part is KtDotQualifiedExpression
-    ) {
-      return true
-    }
-    // this is `Foo` in `com.facebook.Foo`, so everything before it is a package name
-    if (
-        current.text.first().isUpperCase() &&
-            current is KtSimpleNameExpression &&
-            part is KtDotQualifiedExpression
-    ) {
-      return true
-    }
-    // this is the `foo()` in `com.facebook.Foo.foo()` or in `Foo.foo()`
-    if (
-        current is KtCallExpression &&
-            (previous !is KtCallExpression) &&
-            previous.text?.firstOrNull()?.isUpperCase() == true
-    ) {
-      return true
-    }
-    // this is an invocation and the last item, and the previous it not, i.e. `a.b.c()`
-    // keeping it grouped and splitting the arguments makes `a.b(...)` feel like `aab()`
-    return current is KtCallExpression &&
-        previous !is KtCallExpression &&
-        index == parts.indices.last
-  }
-
   override fun visitCallExpression(callExpression: KtCallExpression) {
     builder.sync(callExpression)
     with(callExpression) {
-      visitCallElement(
+      formatFunctionCall(
           calleeExpression,
           typeArgumentList,
           valueArgumentList,
           lambdaArguments,
       )
-    }
-  }
-
-  override fun formatCallElement(
-      callee: KtExpression?,
-      typeArgumentList: KtTypeArgumentList?,
-      argumentList: KtValueArgumentList?,
-      lambdaArguments: List<KtLambdaArgument>,
-      argumentsIndent: Indent,
-      lambdaIndent: Indent,
-      negativeLambdaIndent: Indent,
-  ) {
-    visitCallElement(
-        callee,
-        typeArgumentList,
-        argumentList,
-        lambdaArguments,
-        argumentsIndent,
-        lambdaIndent,
-        negativeLambdaIndent,
-    )
-  }
-
-  /**
-   * Examples `foo<T>(a, b)`, `foo(a)`, `boo()`, `super(a)`
-   *
-   * @param lambdaIndent how to indent [lambdaArguments], if present
-   * @param negativeLambdaIndent the negative indentation of [lambdaIndent]
-   */
-  private fun visitCallElement(
-      callee: KtExpression?,
-      typeArgumentList: KtTypeArgumentList?,
-      argumentList: KtValueArgumentList?,
-      lambdaArguments: List<KtLambdaArgument>,
-      argumentsIndent: Indent = expressionBreakIndent,
-      lambdaIndent: Indent = ZERO,
-      negativeLambdaIndent: Indent = ZERO,
-  ) {
-    // Apply the lambda indent to the callee, type args, value args, and the lambda.
-    // This is undone for the first three by the negative lambda indent.
-    // This way they're in one block, and breaks in the argument list cause a break in the lambda.
-    builder.block(lambdaIndent) {
-
-      // Used to keep track of whether or not we need to indent the lambda
-      // This is based on if there is a break in the argument list
-      var brokeBeforeBrace: BreakTag? = null
-
-      builder.block(negativeLambdaIndent) {
-        visit(callee)
-        builder.block(argumentsIndent) {
-          builder.block(ZERO) { visit(typeArgumentList) }
-          if (argumentList != null) {
-            brokeBeforeBrace = formatValueArgumentList(argumentList)
-          }
-        }
-      }
-      when (lambdaArguments.size) {
-        0 -> {}
-        1 -> {
-          builder.space()
-          visitArgumentInternal(
-              lambdaArguments.single(),
-              wrapInBlock = false,
-              brokeBeforeBrace = brokeBeforeBrace,
-          )
-        }
-        else -> throw ParseError("Maximum one trailing lambda is allowed", lambdaArguments[1])
-      }
-    }
-  }
-
-  /** Example `{ 1 + 1 }` (as lambda) or `{ (x, y) -> x + y }` */
-  override fun visitLambdaExpression(lambdaExpression: KtLambdaExpression) {
-    visitLambdaExpressionInternal(lambdaExpression, brokeBeforeBrace = null)
-  }
-
-  /**
-   * The internal version of [visitLambdaExpression].
-   *
-   * @param brokeBeforeBrace used for tracking if a break was taken right before the lambda
-   *   expression. Useful for scoping functions where we want good looking indentation. For example,
-   *   here we have correct indentation before `bar()` and `car()` because we can detect the break
-   *   after the equals:
-   * ```
-   * fun foo() =
-   *     coroutineScope { x ->
-   *       bar()
-   *       car()
-   *     }
-   * ```
-   */
-  private fun visitLambdaExpressionInternal(
-      lambdaExpression: KtLambdaExpression,
-      brokeBeforeBrace: BreakTag?,
-  ) {
-    builder.sync(lambdaExpression)
-
-    val valueParams = lambdaExpression.valueParameters
-    val hasParams = valueParams.isNotEmpty()
-    val bodyExpression = lambdaExpression.bodyExpression ?: fail()
-    val expressionStatements = bodyExpression.children
-    val hasStatements = expressionStatements.isNotEmpty()
-    val hasComments = bodyExpression.children().any { it is PsiComment }
-    val hasArrow = lambdaExpression.functionLiteral.arrow != null
-
-    fun ifBrokeBeforeBrace(onTrue: Indent, onFalse: Indent): Indent {
-      if (brokeBeforeBrace == null) return onFalse
-      return Indent.If.make(brokeBeforeBrace, onTrue, onFalse)
-    }
-
-    /**
-     * Enable correct formatting of the `fun foo() = scope {` syntax.
-     *
-     * We can't denote the lambda (+ scope function) as a block, since (for multiline lambdas) the
-     * rectangle rule would force the entire lambda onto a lower line. Instead, we conditionally
-     * indent all the interior levels of the lambda based on whether we had to break before the
-     * opening brace (or scope function). This mimics the look of a block when the break is taken.
-     *
-     * These conditional indents should not be used inside interior blocks, since that would apply
-     * the condition twice.
-     */
-    val bracePlusBlockIndent = ifBrokeBeforeBrace(blockPlusExpressionBreakIndent, blockIndent)
-    val bracePlusExpressionIndent =
-        ifBrokeBeforeBrace(doubleExpressionBreakIndent, expressionBreakIndent)
-    val bracePlusZeroIndent = ifBrokeBeforeBrace(expressionBreakIndent, ZERO)
-
-    builder.token("{")
-
-    if (hasParams || hasArrow) {
-      builder.space()
-      builder.block(bracePlusExpressionIndent) { formatCommaSeparatedList(valueParams) }
-      builder.block(bracePlusBlockIndent) {
-        if (lambdaExpression.functionLiteral.valueParameterList?.trailingComma != null) {
-          builder.token(",")
-          builder.forcedBreak()
-        } else if (hasParams) {
-          builder.breakOp(Doc.FillMode.INDEPENDENT, " ", ZERO)
-        }
-        builder.token("->")
-      }
-    }
-
-    if (hasParams || hasArrow || hasStatements || hasComments) {
-      builder.breakOp(Doc.FillMode.UNIFIED, " ", bracePlusZeroIndent)
-    }
-
-    if (hasStatements) {
-      builder.breakOp(Doc.FillMode.UNIFIED, "", bracePlusBlockIndent)
-      builder.block(bracePlusBlockIndent) {
-        builder.blankLineWanted(OpsBuilder.BlankLineWanted.NO)
-
-        val shouldForceMultiline =
-            options.preserveLambdaBreaks && lambdaExpression.hasSourceNewlineInLambdaBody
-
-        if (
-            !shouldForceMultiline &&
-                expressionStatements.size == 1 &&
-                expressionStatements.first() !is KtReturnExpression &&
-                !bodyExpression.startsWithComment()
-        ) {
-          visitStatement(expressionStatements[0])
-        } else {
-          visitStatements(expressionStatements)
-        }
-        builder.breakOp(Doc.FillMode.UNIFIED, " ", bracePlusZeroIndent)
-      }
-    } else if (hasComments) {
-      val blockComments =
-          bodyExpression.children().filter { it is PsiComment && it.text.startsWith("/*") }.toList()
-      builder.breakOp(Doc.FillMode.UNIFIED, "", bracePlusBlockIndent)
-      builder.block(bracePlusBlockIndent) {
-        builder.fenceComments()
-        builder.blankLineWanted(OpsBuilder.BlankLineWanted.NO)
-        if (blockComments.size == 1) {
-          builder.token(blockComments[0].text)
-        } else {
-          for ((i, comment) in blockComments.withIndex()) {
-            if (i > 0) {
-              builder.forcedBreak()
-            }
-            builder.token(comment.text)
-          }
-        }
-        builder.breakOp(Doc.FillMode.UNIFIED, " ", bracePlusZeroIndent)
-      }
-    }
-
-    if (hasParams || hasArrow || hasStatements || hasComments) {
-      // If we had to break in the body, ensure there is a break before the closing brace
-      builder.breakOp(Doc.FillMode.UNIFIED, "", bracePlusZeroIndent)
-    }
-    builder.block(bracePlusZeroIndent) {
-      builder.fenceComments()
-      builder.token("}", blockIndent)
     }
   }
 
@@ -858,55 +367,6 @@ open class KotlinInputAstVisitor(
         if (expression.text.isNotEmpty()) {
           builder.token(expression.text)
         }
-      }
-    }
-  }
-
-  /** Example `a` in `foo(a)`, or `*a`, or `limit = 50` */
-  override fun visitArgument(argument: KtValueArgument) {
-    visitArgumentInternal(
-        argument,
-        wrapInBlock = true,
-        brokeBeforeBrace = null,
-    )
-  }
-
-  /**
-   * The internal version of [visitArgument].
-   *
-   * @param wrapInBlock if true places the argument expression in a block.
-   */
-  private fun visitArgumentInternal(
-      argument: KtValueArgument,
-      wrapInBlock: Boolean,
-      brokeBeforeBrace: BreakTag?,
-  ) {
-    builder.sync(argument)
-    val hasArgName = argument.getArgumentName() != null
-    val isLambda = argument.getArgumentExpression() is KtLambdaExpression
-    if (hasArgName) {
-      visit(argument.getArgumentName())
-      builder.space()
-      builder.token("=")
-      if (isLambda) {
-        builder.space()
-      }
-    }
-    val indent = if (hasArgName && !isLambda) expressionBreakIndent else ZERO
-    builder.block(indent, isEnabled = wrapInBlock) {
-      if (hasArgName && !isLambda) {
-        builder.breakOp(Doc.FillMode.INDEPENDENT, " ", ZERO)
-      }
-      if (argument.isSpread) {
-        builder.token("*")
-      }
-      if (isLambda) {
-        visitLambdaExpressionInternal(
-            argument.getArgumentExpression() as KtLambdaExpression,
-            brokeBeforeBrace = brokeBeforeBrace,
-        )
-      } else {
-        visit(argument.getArgumentExpression())
       }
     }
   }
@@ -944,7 +404,7 @@ open class KotlinInputAstVisitor(
       visit(expression.left)
       builder.space()
       builder.token(expression.operationReference.text)
-      visitLambdaOrScopingFunction(expression.right)
+      formatLambdaOrScopingFunction(expression.right)
       return
     }
 
@@ -1124,12 +584,12 @@ open class KotlinInputAstVisitor(
           builder.space()
           visit(delegate)
         } else if (delegateExpr != null && delegateExpr.isChainedScopingFunction) {
-          visitChainedScopingFunction(delegateExpr, emitLeadingBreak = true)
+          formatChainedScopingFunction(delegateExpr, emitLeadingBreak = true)
         } else if (delegateExpr.isBlockLikeCall) {
           builder.space()
           visit(delegate)
         } else if (delegateExpr != null && delegateExpr.isChainedBlockLikeCall) {
-          visitChainedBlockLikeCall(delegateExpr, emitLeadingBreak = true)
+          formatChainedBlockLikeCall(delegateExpr, emitLeadingBreak = true)
         } else {
           builder.breakOp(Doc.FillMode.UNIFIED, " ", expressionBreakIndent)
           builder.block(expressionBreakIndent) {
@@ -1243,125 +703,6 @@ open class KotlinInputAstVisitor(
     }
   }
 
-  override fun formatChainedBlockLikeCall(
-      expression: KtQualifiedExpression,
-      emitLeadingBreak: Boolean,
-  ) {
-    visitChainedBlockLikeCall(expression, emitLeadingBreak)
-  }
-
-  private fun visitChainedBlockLikeCall(
-      expression: KtQualifiedExpression,
-      emitLeadingBreak: Boolean,
-  ) {
-    val parts = expression.chainParts
-    if (emitLeadingBreak) {
-      builder.space()
-    }
-    visit(parts[0])
-
-    builder.block(expressionBreakIndent) {
-      for (i in 1 until parts.size) {
-        val part = parts[i] as KtQualifiedExpression
-        builder.forcedBreak()
-        builder.token(part.operationSign.value)
-        val selectorExpression = part.selectorExpression
-        if (selectorExpression is KtCallExpression) {
-          visit(selectorExpression.calleeExpression)
-          visitCallElement(
-              null,
-              selectorExpression.typeArgumentList,
-              selectorExpression.valueArgumentList,
-              selectorExpression.lambdaArguments,
-          )
-        } else {
-          visit(selectorExpression)
-        }
-      }
-    }
-  }
-
-  override fun formatChainedScopingFunction(
-      expression: KtQualifiedExpression,
-      emitLeadingBreak: Boolean,
-  ) {
-    visitChainedScopingFunction(expression, emitLeadingBreak)
-  }
-
-  private fun visitChainedScopingFunction(
-      expression: KtQualifiedExpression,
-      emitLeadingBreak: Boolean,
-  ) {
-    val parts = expression.chainParts
-    val root = parts[0]
-    val forceBreakBeforeChain = root.isMultilineScopingFunction
-
-    visitLambdaOrScopingFunction(root, emitLeadingBreak = emitLeadingBreak)
-
-    // The break before each selector must stay outside the block below, at the same level as
-    // the lambda, so that it is taken exactly when the lambda breaks. Inside the block it
-    // would fire only when the selector itself is too long, so a lambda broken by max width
-    // would keep its selector on the closing brace's line — and the next format pass, seeing
-    // a multiline lambda in the source, would force the selector onto its own line (#640).
-    val fillMode = if (forceBreakBeforeChain) Doc.FillMode.FORCED else Doc.FillMode.UNIFIED
-    for (i in 1 until parts.size) {
-      val part = parts[i] as KtQualifiedExpression
-      builder.breakOp(fillMode, "", expressionBreakIndent)
-      builder.block(expressionBreakIndent) {
-        builder.token(part.operationSign.value)
-        val selectorExpression = part.selectorExpression
-        if (selectorExpression is KtCallExpression) {
-          visit(selectorExpression.calleeExpression)
-          visitCallElement(
-              null,
-              selectorExpression.typeArgumentList,
-              selectorExpression.valueArgumentList,
-              selectorExpression.lambdaArguments,
-          )
-        } else {
-          visit(selectorExpression)
-        }
-      }
-    }
-  }
-
-  override fun formatLambdaOrScopingFunction(expr: PsiElement?, emitLeadingBreak: Boolean) {
-    visitLambdaOrScopingFunction(expr, emitLeadingBreak)
-  }
-
-  private fun visitLambdaOrScopingFunction(expr: PsiElement?, emitLeadingBreak: Boolean = true) {
-    val breakToExpr = genSym()
-    val breakSpace = if (emitLeadingBreak) " " else ""
-    builder.breakOp(
-        Doc.FillMode.INDEPENDENT,
-        breakSpace,
-        expressionBreakIndent,
-        Optional.of(breakToExpr),
-    )
-
-    var carry = expr
-    if (carry is KtQualifiedExpression && carry.receiverExpression is KtSimpleNameExpression) {
-      visit(carry.receiverExpression)
-      builder.token(carry.operationSign.value)
-      carry = carry.selectorExpression
-    }
-    if (carry is KtCallExpression) {
-      visit(carry.calleeExpression)
-      builder.space()
-      carry = carry.lambdaArguments[0].getArgumentExpression()
-    }
-    if (carry is KtLabeledExpression) {
-      visit(carry.labelQualifier)
-      carry = carry.baseExpression ?: fail()
-    }
-    if (carry is KtLambdaExpression) {
-      visitLambdaExpressionInternal(carry, brokeBeforeBrace = breakToExpr)
-      return
-    }
-
-    throw AssertionError(carry)
-  }
-
   override fun visitClassOrObject(classOrObject: KtClassOrObject) {
     builder.sync(classOrObject)
     val contextReceiverList =
@@ -1459,7 +800,7 @@ open class KotlinInputAstVisitor(
     // visitReferenceExpression, but calls visitElement instead.
     builder.block(ZERO) {
       builder.token(if (call.isCallToThis) "this" else "super")
-      visitCallElement(
+      formatFunctionCall(
           null,
           call.typeArgumentList,
           call.valueArgumentList,
@@ -1542,7 +883,7 @@ open class KotlinInputAstVisitor(
 
   override fun visitSuperTypeCallEntry(call: KtSuperTypeCallEntry) {
     builder.sync(call)
-    visitCallElement(call.calleeExpression, null, call.valueArgumentList, call.lambdaArguments)
+    formatFunctionCall(call.calleeExpression, null, call.valueArgumentList, call.lambdaArguments)
   }
 
   /**
@@ -1674,7 +1015,7 @@ open class KotlinInputAstVisitor(
 
   override fun visitBlockExpression(expression: KtBlockExpression) {
     builder.sync(expression)
-    emitBracedBlock(expression) { children -> visitStatements(children) }
+    emitBracedBlock(expression) { children -> formatStatements(children) }
   }
 
   override fun visitWhenConditionWithExpression(condition: KtWhenConditionWithExpression) {
@@ -1734,34 +1075,6 @@ open class KotlinInputAstVisitor(
           }
         }
       }
-    }
-  }
-
-  /** Example `a[3]`, `b["a", 5]` or `a.b.c[4]` */
-  override fun visitArrayAccessExpression(expression: KtArrayAccessExpression) {
-    builder.sync(expression)
-    if (expression.arrayExpression is KtQualifiedExpression) {
-      emitQualifiedExpression(expression)
-    } else {
-      visit(expression.arrayExpression)
-      visitArrayAccessBrackets(expression)
-    }
-  }
-
-  /**
-   * Example `[3]` in `a[3]` or `a[3].b` Separated since it needs to be used from a top level array
-   * expression (`a[3]`) and from within a qualified chain (`a[3].b)
-   */
-  private fun visitArrayAccessBrackets(expression: KtArrayAccessExpression) {
-    builder.block(expressionBreakIndent) {
-      formatCommaSeparatedList(
-          expression.indexExpressions,
-          forceMultiline = expression.trailingComma != null,
-          wrapInBlock = true,
-          prefix = "[",
-          postfix = "]",
-          breakBeforePostfix = false,
-      )
     }
   }
 
@@ -1938,7 +1251,7 @@ open class KotlinInputAstVisitor(
     builder.sync(expression)
     val receiverExpression = expression.receiverExpression
     if (receiverExpression is KtCallExpression) {
-      visitCallElement(
+      formatFunctionCall(
           receiverExpression.calleeExpression,
           receiverExpression.typeArgumentList,
           receiverExpression.valueArgumentList,
@@ -2108,16 +1421,6 @@ open class KotlinInputAstVisitor(
       inExpression.removeLast()
     }
     builder.checkClosed(previous)
-  }
-
-  /**
-   * Throws a formatting error
-   *
-   * This is used as `expr ?: fail()` to avoid using the !! operator and provide better error
-   * messages.
-   */
-  private fun fail(message: String = "Unexpected"): Nothing {
-    throw FormattingError(builder.diagnostic(message))
   }
 
   /** Helper function to improve readability */
