+++ core/src/main/kotlin/org/jetbrains/ktfmt/format/visitor/FileFormatter.kt
@@ -1,7 +1,9 @@
 package org.jetbrains.ktfmt.format.visitor
 
+import com.google.googlejavaformat.Indent.Const.ZERO
 import com.google.googlejavaformat.OpsBuilder
 import org.jetbrains.kotlin.com.intellij.psi.PsiComment
+import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.psi.KtElement
 import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.KtProperty
@@ -48,5 +50,25 @@ interface FileFormatter : KotlinAstFormatter {
       first = false
     }
     markForPartialFormat()
+  }
+
+  override fun formatStatement(statement: PsiElement) {
+    builder.block(ZERO) { format(statement) }
+    builder.guessToken(";")
+  }
+
+  override fun formatStatements(statements: Array<PsiElement>) {
+    var first = true
+    builder.guessToken(";")
+    for (statement in statements) {
+      builder.forcedBreak()
+      if (!first) {
+        builder.blankLineWanted(OpsBuilder.BlankLineWanted.PRESERVE)
+      }
+      first = false
+      markForPartialFormat()
+      formatStatement(statement)
+      markForPartialFormat()
+    }
   }
 }
